Redesigns evaluateRegex method
This is the first in what I think will be a series of changes.
I plan on changing all three of the methods that filter lists to have then filter only strings.
Then I can call them in a single stream from the main method.
I think it will be easier to debug!